requirements.txt
beautifulsoup4==4.12.2
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
lxml==5.1.0
fastapi==0.109.2
uvicorn==0.27.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.7
starlette==0.36.3